pyproject.toml
[build-system]
requires = [ "poetry>=0.12",]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "opsiclientd"
version = "4.3.3.0"
description = "opsi client daemon. The service running on opsi clients."
maintainers = [ "uib GmbH <info@uib.de>",]
authors = [ "uib GmbH <info@uib.de>",]
homepage = "https://www.opsi.org"
license = "AGPL-3.0"
include = [ "run-opsiclientd", "opsiclientd_data/**/*",]
[[tool.poetry.source]]
name = "uibpypi"
url = "https://pypi.uib.gmbh/simple"
priority = "primary"

[[tool.poetry.source]]
name = "PyPI"
priority = "supplemental"

[tool.ruff]
line-length = 140

[tool.poetry.dependencies]
python = "^3.11,<3.12"
autobahn = "^22.4"
macholib = "^1.14"
msgpack = "^1.0"
netifaces = "^0.11"
pyelftools = "^0.29"
python-opsi = "^4.3.0.13"
python-opsi-common = "^4.3.4"
zeroconf = "^0.131"
fastapi = "^0.110"
uvicorn = "^0.29"
click = "^8.1"
h11 = "^0.14"
websockets = "^12.0"
starlette = "^0.37.2"

[tool.poetry.scripts]
opsiclientd = "opsiclientd.__main__:opsiclientd"
opsiclientd_rpc = "opsiclientd.__main__:opsiclientd_rpc"
action_processor_starter = "opsiclientd.__main__:action_processor_starter"

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
follow_imports = "silent"

[tool.ruff.lint]
ignore = ["E402",]

[tool.ruff.format]
indent-style = "tab"

[tool.poetry.dependencies.distro]
platform = "linux"
version = "^1.5"

[tool.poetry.dependencies.websocket-client]
git = "https://github.com/j-a-n/websocket-client.git"
branch = "fix-get_proxy_info-for-https"

[tool.poetry.dependencies.pywin32]
platform = "win32"
version = ">=303,<304"

[tool.poetry.dependencies.pywinpty]
platform = "win32"
version = "^2.0"

[tool.poetry.dependencies.wmi]
platform = "win32"
version = "^1.5"

[tool.poetry.dev-dependencies.macholib]
version = "^1.14"
platform = "darwin"

[tool.poetry.dev-dependencies.pefile]
version = "^2023.2"
platform = "win32"

[tool.poetry.dev-dependencies.pywin32-ctypes]
version = "^0.2"
platform = "win32"

[tool.poetry.group.dev.dependencies]
mock = "^5.0"
mypy = "^1.4"
pytest = "^8.0"
pytest-cov = "^4.0"
ruff = "^0.3"
pyinstaller = "5.13"
types-pywin32 = "^306.0"
types-psutil = "^5.9"
types-pyopenssl = "^24.0"
types-mock = "^5.1"
types-requests = "^2.31"
httpx = "^0.27.0"
